pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi==0.104.1",
    "google-api-python-client>=2.190.0",
    "google-auth==2.23.3",
    "google-cloud-bigquery==3.13.0",
    "google-cloud-logging==3.8.0",
    "google-generativeai>=0.8.6",
    "numpy>=2.4.2",
    "pillow-heif>=1.3.0",
    "pymupdf>=1.27.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "uvicorn==0.24.0",
    "websockets>=16.0",
]
